airflow/requirements.txt
apache-airflow==2.10.5
apache-airflow-providers-google==15.1.0
pandas
requests
pytz
pyarrow
google-cloud-bigquery==3.34.0
